% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_textual.r
\name{display_<text>}
\alias{display_<text>}
\alias{display_html}
\alias{display_javascript}
\alias{display_json}
\alias{display_latex}
\alias{display_markdown}
\title{Display a specific textual output}
\usage{
display_json(data = NULL, file = NULL)

display_javascript(data = NULL, file = NULL)

display_html(data = NULL, file = NULL)

display_markdown(data = NULL, file = NULL)

display_latex(data = NULL, file = NULL)
}
\arguments{
\item{data}{The code or markup content as a \code{\link[base]{character}} vector}

\item{file}{The path to a file or a \code{\link[base]{connection}} containing the content}
}
\description{
Either \code{data} or \code{file} must be passed.
}
\seealso{
\code{\link{display_<image>}}
}

