% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/display.R
\name{display_alternatives}
\alias{display_alternatives}
\title{Display a set of formatted objects that the Jupyter frontend/output can select from}
\usage{
display_alternatives(..., metadata = namedlist())
}
\arguments{
\item{...}{Mimetypes and content.

You can either give raw vectors or character vectors as content.

Also you can give a list of which the first element will be content, the rest metadata.}

\item{metadata}{Initial metadata. Mimetype-wise data will be added}
}
\description{
Display a set of formatted objects that the Jupyter frontend/output can select from
}
\examples{
\dontrun{
display_alternatives(
  'text/html'       = '<em>Wop</em>'
  'application/pdf' = base64encode('test.pdf'),
  'image/png'       = list(some.png.data, width = 800, height = 600),
  'image/svg+xml'   = '<svg></svg>')
}
}

