% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display_html}
\alias{display_html}
\title{Display HTML output}
\usage{
display_html(data = NULL, file = NULL)
}
\arguments{
\item{data}{The HTML code as a character vector}

\item{file}{The path to a HTML file or a connection}
}
\description{
Either data or file must be passed.
}

