% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/display.R
\name{display_javascript}
\alias{display_javascript}
\title{Display (execute) Javascript}
\usage{
display_javascript(data = NULL, file = NULL)
}
\arguments{
\item{data}{The Javascript code as a character vector}

\item{file}{The path to a Javascript file or a connection}
}
\description{
Either data or file must be passed.
}

