% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/display.R
\name{display_jpeg}
\alias{display_jpeg}
\title{Display JPEG output}
\usage{
display_jpeg(data = NULL, file = NULL, width = NULL, height = NULL)
}
\arguments{
\item{data}{The JPEG data as a raw vector}

\item{file}{The path to a JPEG file or a connection}

\item{width}{The width to display the image}

\item{height}{The height to display the image}
}
\description{
Either data or file must be passed.
}

