% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_textual.r
\name{display_json}
\alias{display_json}
\title{Display JSON}
\usage{
display_json(data = NULL, file = NULL)
}
\arguments{
\item{data}{The JSON code as a character vector}

\item{file}{The path to a JSON file or a connection}
}
\description{
Either data or file must be passed.
}

