% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/display.R
\name{display_latex}
\alias{display_latex}
\title{Display LaTeX output}
\usage{
display_latex(data = NULL, file = NULL)
}
\arguments{
\item{data}{The LaTeX code as a character vector}

\item{file}{The path to a LaTeX file or a connection}
}
\description{
Either data or file must be passed.
}

