% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display_markdown}
\alias{display_markdown}
\title{Display Markdown output}
\usage{
display_markdown(data = NULL, file = NULL)
}
\arguments{
\item{data}{The Markdown code as a character vector}

\item{file}{The path to a Markdown file or a connection}
}
\description{
Either data or file must be passed.
}

