% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_images.r
\name{display_pdf}
\alias{display_pdf}
\title{Display PDF output}
\usage{
display_pdf(data = NULL, file = NULL, width = NULL, height = NULL)
}
\arguments{
\item{data}{The PDF data as a raw vector}

\item{file}{The path to a PDF file or a connection}

\item{width}{The width to display the document}

\item{height}{The height to display the document}
}
\description{
Either data or file must be passed.
}

