\name{display_png}
\alias{display_png}
\title{Display PNG output}
\usage{
display_png(data = NULL, filename = NULL, width = NULL, height = NULL)
}
\arguments{
  \item{data}{The PNG data as a raw vector}

  \item{filename}{The path to a PNG file}

  \item{width}{The width to display the image}

  \item{height}{The height to display the image}
}
\description{
Either data or filename must be passed.
}

