% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_images.r
\name{display_svg}
\alias{display_svg}
\title{Display SVG output}
\usage{
display_svg(data = NULL, file = NULL, width = NULL, height = NULL)
}
\arguments{
\item{data}{The SVG data as character vector}

\item{file}{The path to a SVG file or a connection}

\item{width}{The width to display the image}

\item{height}{The height to display the image}
}
\description{
Either data or file must be passed.
}

