% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.r
\docType{data}
\name{IRdisplay-options}
\alias{IRdisplay-options}
\alias{irdisplay_option_defaults}
\title{IRdisplay options}
\format{An object of class \code{list} of length 2.}
\usage{
irdisplay_option_defaults
}
\description{
Some \link{options} to control the formats \code{\link{display}} and \code{\link{prepare_mimebundle}} emit,
and the function they use to display them.
}
\section{Options}{


\describe{

\item{\code{jupyter.display_mimetypes}}{
 The default is all MIME types supported by Jupyter.
}
\item{\code{jupyter.base_display_func}}{
	Function used by \code{\link{display}} and all \code{\link{display_<text>}} / \code{\link{display_<image>}} functions.
	Has the signature \code{function(data, metadata = NULL)}.
	Per default emits a \code{\link{warning}}, and is set when running an \code{IRkernel}.
}

}
}
\keyword{datasets}

