% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.r
\name{prepare_mimebundle}
\alias{prepare_mimebundle}
\title{Create a mimebundle for multiple reprs}
\usage{
prepare_mimebundle(obj, mimetypes = getOption("jupyter.display_mimetypes"),
  metadata = NULL, error_handler = stop)
}
\arguments{
\item{obj}{The object to create representations for displayed}

\item{mimetypes}{Mimetypes to create reprs for}

\item{metadata}{Metadata to attach to the result (can be expanded by additional metadata)}

\item{error_handler}{Function used when errors in individual reprs occur}
}
\value{
A list with the two items \code{data} (a list mapping mimetype to character) and \code{metadata} (mapping mimetype to lists of metadata)
}
\description{
Create a mimebundle for multiple reprs
}

