\name{display}
\alias{display}
\title{Display data by mimetype}
\usage{
display(mimetype, content, metadata)
}
\arguments{
  \item{mimetype}{The mimetype of the data}

  \item{content}{The data as a length 1 string vector}

  \item{metadata}{A named list of metadata}
}
\description{
Display data by mimetype
}

