% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/display.R
\name{display}
\alias{display}
\title{Display data by mimetype, with optional alternative representations.}
\usage{
display(data, metadata = NULL)
}
\arguments{
\item{data}{A named list mapping mimetypes to content (base64 encoded for binary data)}

\item{metadata}{A named list mapping mimetypes to named lists of specific metadata}
}
\description{
Display data by mimetype, with optional alternative representations.
}

