% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.r
\name{display}
\alias{display}
\alias{prepare_mimebundle}
\title{Create and use multiple available reprs}
\usage{
display(obj)

prepare_mimebundle(obj, mimetypes = getOption("jupyter.display_mimetypes"),
  metadata = NULL, error_handler = stop)
}
\arguments{
\item{obj}{The object to create representations for}

\item{mimetypes}{Mimetypes to create reprs for. The defaults are defined by the option \code{jupyter.display_mimetypes}. (see: \link{IRdisplay-options})}

\item{metadata}{Metadata to attach to the result (can be expanded by additional metadata)}

\item{error_handler}{Function used when errors in individual reprs occur}
}
\value{
\code{prepare_mimebundle} returns a list with items corresponding to the parameters of \code{\link{publish_mimebundle}} (\code{data} and \code{metadata})
}
\description{
Both functions create a mimebundle for multiple reprs.
\code{display} proceeds to publish it using \code{\link{publish_mimebundle}}.
\code{prepare_mimebundle} returns it (see \emph{Value} for details)
}
\examples{
bundle <- prepare_mimebundle(diag(3))

\dontrun{## (Run inside of an IRkernel)
display(help(display))}

}
\seealso{
\code{\link{publish_mimebundle}}
}

