% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display}
\alias{display}
\title{Display an object using any available reprs}
\usage{
display(obj)
}
\arguments{
\item{obj}{The object to be displayed}
}
\description{
Display an object using any available reprs
}

