\name{display1}
\alias{display1}
\title{Display a single type of data}
\usage{
display1(mimetype, content, metadata = NULL)
}
\arguments{
  \item{mimetype}{The mimetype of the data}

  \item{content}{The data as a length 1 string vector}

  \item{metadata}{A named list of metadata}
}
\description{
Display a single type of data
}

