% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display_pdf}
\alias{display_pdf}
\title{Display PDF output}
\usage{
display_pdf(data = NULL, file = NULL)
}
\arguments{
\item{data}{The PDF data as a raw vector}

\item{file}{The path to a PDF file or a connection}
}
\description{
Either data or file must be passed.
}

