% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_images.r
\name{display_png}
\alias{display_png}
\title{Display PNG output}
\usage{
display_png(data = NULL, file = NULL, width = NULL, height = NULL)
}
\arguments{
\item{data}{The PNG data as a raw vector}

\item{file}{The path to a PNG file or a connection}

\item{width}{The width to display the image}

\item{height}{The height to display the image}
}
\description{
Either data or file must be passed.
}

