% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/display.R
\name{display_svg}
\alias{display_svg}
\title{Display SVG output}
\usage{
display_svg(data = NULL, file = NULL, width = NULL, height = NULL)
}
\arguments{
\item{data}{The SVG data as character vector}

\item{width}{The width to display the image}

\item{height}{The height to display the image}

\item{filename}{The path to a SVG file}
}
\description{
Either data or filename must be passed.
}

