% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/display.R
\name{display_svg}
\alias{display_svg}
\title{Display SVG output}
\usage{
display_svg(data = NULL, file = NULL)
}
\arguments{
\item{data}{The SVG data as character vector}

\item{filename}{The path to a SVG file}
}
\description{
Either data or file must be passed.
}

