% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.r
\name{publish_mimebundle}
\alias{publish_mimebundle}
\title{Display data by mimetype, with optional alternative representations.}
\usage{
publish_mimebundle(data, metadata = NULL)
}
\arguments{
\item{data}{A named list mapping mimetypes to content (\code{\link[base]{character}} or \code{\link[base]{raw} vectors})}

\item{metadata}{A named list mapping mimetypes to named lists of metadata, e.g. \code{list('image/png' = list(width = 5))}}
}
\description{
Calls the function stored as option value of \code{jupyter.base_display_func}.
}
\seealso{
\code{\link{prepare_mimebundle}}
}

